+++ src/com/reason/lang/core/type/ORTypes.java
@@ -3,26 +3,26 @@ package com.reason.lang.core.type;
 public abstract class ORTypes {
   // Stubbed elements
 
-  public ORCompositeType C_FAKE_MODULE;
+  public ORCompositeType C_CLASS_DECLARATION;
   public ORCompositeType C_EXCEPTION_DECLARATION;
-  public ORCompositeType C_TYPE_DECLARATION;
   public ORCompositeType C_EXTERNAL_DECLARATION;
-  public ORCompositeType C_LET_DECLARATION;
-  public ORCompositeType C_MODULE_DECLARATION;
-  public ORCompositeType C_VAL_DECLARATION;
+  public ORCompositeType C_FAKE_MODULE;
   public ORCompositeType C_FUN_PARAM;
   public ORCompositeType C_FUNCTOR_DECLARATION;
   public ORCompositeType C_FUNCTOR_PARAM;
-  public ORCompositeType C_RECORD_FIELD;
-  public ORCompositeType C_VARIANT_DECLARATION;
   public ORCompositeType C_INCLUDE;
+  public ORCompositeType C_LET_DECLARATION;
+  public ORCompositeType C_MODULE_DECLARATION;
+  public ORCompositeType C_RECORD_FIELD;
   public ORCompositeType C_OPEN;
+  public ORCompositeType C_TYPE_DECLARATION;
+  public ORCompositeType C_VAL_DECLARATION;
+  public ORCompositeType C_VARIANT_DECLARATION;
 
   // Composite element types
 
   public ORCompositeType C_ANNOTATION;
   public ORCompositeType C_ASSERT_STMT;
-  public ORCompositeType C_CLASS_DECLARATION;
   public ORCompositeType C_CLASS_CONSTR;
   public ORCompositeType C_CLASS_FIELD;
   public ORCompositeType C_CLASS_METHOD;