+++ tests/com/reason/lang/BaseParsingTestCase.java
@@ -7,7 +7,7 @@ import com.intellij.psi.util.*;
 import com.intellij.testFramework.*;
 import com.reason.ide.files.*;
 import com.reason.lang.core.*;
-import com.reason.lang.core.psi.PsiClass;
+import com.reason.lang.core.psi.PsiKlass;
 import com.reason.lang.core.psi.PsiType;
 import com.reason.lang.core.psi.*;
 import com.reason.lang.core.psi.impl.*;
@@ -61,7 +61,7 @@ public abstract class BaseParsingTestCase extends ParsingTestCase {
     }
 
     @NotNull
-    protected Collection<PsiClass> classExpressions(@NotNull PsiFile file) {
+    protected Collection<PsiKlass> classExpressions(@NotNull PsiFile file) {
         return PsiFileHelper.getClassExpressions(file);
     }